+++ src/editortoolkit_neume.cpp
@@ -77,8 +77,14 @@ bool EditorToolkitNeume::ParseEditorAction(const std::string &json_editorAction)
         m_editInfo.import("message", "'param' can only be an array for a chain action.");
         return false;
     }
-
-    if (action == "drag") {
+    if (action == "addSyl") {
+        std::string elementId, sylText;
+        if (this->ParseAddSylAction(json.get<jsonxx::Object>("param"), &elementId, &sylText)) {
+            return this->AddSyl(elementId, sylText);
+        }
+        LogWarning("Could not parse the addSyl action");
+    }
+    else if (action == "drag") {
         std::string elementId;
         int x, y;
         if (this->ParseDragAction(json.get<jsonxx::Object>("param"), &elementId, &x, &y)) {
@@ -277,6 +283,60 @@ bool EditorToolkitNeume::Chain(jsonxx::Array actions)
     return status;
 }
 
+bool EditorToolkitNeume::AddSyl(std::string elementId, std::string sylText)
+{
+    if (!m_doc->GetDrawingPage()) {
+        LogError("Could not get the drawing page.");
+        m_editInfo.import("status", "FAILURE");
+        m_editInfo.import("message", "Could not get the drawing page.");
+        return true;
+    }
+
+    Syllable *syllable = dynamic_cast<Syllable *>(m_doc->GetDrawingPage()->FindDescendantByID(elementId));
+    if (syllable == NULL) {
+        LogError("Unable to find syllable with id %s", elementId.c_str());
+        m_editInfo.import("status", "FAILURE");
+        m_editInfo.import("message", "Unable to find neume with id " + elementId + ".");
+        return false;
+    }
+
+    // Create a new syl element
+    Syl *syl = new Syl();
+    Text *text = new Text();
+    text->SetText(UTF8to32(sylText));
+    syl->AddChild(text);
+    syllable->AddChild(syl);
+
+    // Create default bounding box if facs
+    if (m_doc->HasFacsimile()) {
+        Zone *zone = new Zone();
+        Staff *staff = syllable->GetAncestorStaff();
+        const int staffSize = m_doc->GetDrawingDoubleUnit(staff->m_drawingStaffSize);
+
+        zone->SetUlx(syllable->GetFirst(NEUME)->GetFirst(NC)->GetFacsimileInterface()->GetZone()->GetUlx());
+        zone->SetUly(staff->GetFacsimileInterface()->GetZone()->GetLry());
+        zone->SetLrx(syllable->GetLast(NEUME)->GetLast(NC)->GetFacsimileInterface()->GetZone()->GetLrx());
+        zone->SetLry(zone->GetUly() + staffSize * 2);
+
+        // Make bbox larger if it has less than 2 ncs
+        if (syllable->GetChildCount(NC, 2) <= 2) {
+            zone->SetLrx(zone->GetLrx() + 50);
+        }
+
+        assert(m_doc->GetFacsimile());
+        m_doc->GetFacsimile()->FindDescendantByType(SURFACE)->AddChild(zone);
+        FacsimileInterface *fi = syl->GetFacsimileInterface();
+        assert(fi);
+        fi->AttachZone(zone);
+
+        if (m_doc->IsTranscription() && m_doc->HasFacsimile()) m_doc->SyncFromFacsimileDoc();
+    }
+
+    m_editInfo.import("uuid", elementId);
+    m_editInfo.import("status", "OK");
+    m_editInfo.import("message", "");
+    return true;
+}
 //////////////////////////////////////////////////////////////////////////////////////////////////
 // Author: jacob-hutnyk
 //
@@ -360,7 +420,7 @@ bool EditorToolkitNeume::ClefMovementHandler(Clef *clef, int x, int y)
 
     // Note that y param is relative to initial position for clefs
     int initialClefLine = clef->GetLine();
-    int clefLine = round(((double)y - x * tan(staff->GetDrawingRotate() * M_PI / 180.0))
+    int clefLine = round(((double)y - x * tan(staff->GetDrawingRotation() * M_PI / 180.0))
             / (double)m_doc->GetDrawingDoubleUnit(staff->m_drawingStaffSize)
         + initialClefLine);
 
@@ -384,7 +444,7 @@ bool EditorToolkitNeume::ClefMovementHandler(Clef *clef, int x, int y)
         Zone *zone = clef->GetZone();
         assert(zone);
         int y = (clefLine - initialClefLine) * 2 * staff->m_drawingStaffSize
-            - x * tan(staff->GetDrawingRotate() * M_PI / 180.0);
+            - x * tan(staff->GetDrawingRotation() * M_PI / 180.0);
         zone->ShiftByXY(x, -y);
     }
 
@@ -875,7 +935,7 @@ bool EditorToolkitNeume::Insert(std::string elementType, std::string staffId, in
             int bboxOffsetX = 50;
 
             // calculate staff rotation offset
-            double theta = staff->GetDrawingRotate();
+            double theta = staff->GetDrawingRotation();
             int offsetY = 0;
             if (theta) {
                 double factor = 1.3;
@@ -967,7 +1027,7 @@ bool EditorToolkitNeume::Insert(std::string elementType, std::string staffId, in
                 }
 
                 // Apply offset due to rotate
-                newUly += (newUlx - ulx) * tan(-staff->GetDrawingRotate() * M_PI / 180.0);
+                newUly += (newUlx - ulx) * tan(-staff->GetDrawingRotation() * M_PI / 180.0);
                 newZone->SetUlx(newUlx - offsetX);
                 newZone->SetUly(newUly);
                 newZone->SetLrx(newUlx + offsetX);
@@ -1030,7 +1090,7 @@ bool EditorToolkitNeume::Insert(std::string elementType, std::string staffId, in
         clef->SetShape(clefShape);
         int yDiff = -staff->GetZone()->GetUly() + uly;
         yDiff += ((ulx - staff->GetZone()->GetUlx()))
-            * tan(-staff->GetDrawingRotate() * M_PI / 180.0); // Subtract distance due to rotate.
+            * tan(-staff->GetDrawingRotation() * M_PI / 180.0); // Subtract distance due to rotate.
         int clefLine = staff->m_drawingLines - round((double)yDiff / (double)staffSize);
         clef->SetLine(clefLine);
 
@@ -1083,15 +1143,11 @@ bool EditorToolkitNeume::Insert(std::string elementType, std::string staffId, in
             = (int)(m_doc->GetDrawingDoubleUnit(staff->m_drawingStaffSize) / NOTE_HEIGHT_TO_STAFF_SIZE_RATIO);
         const int noteWidth
             = (int)(m_doc->GetDrawingDoubleUnit(staff->m_drawingStaffSize) / NOTE_WIDTH_TO_STAFF_SIZE_RATIO);
-        const int offsetX = (int)(noteWidth / 4);
 
-        ulx -= noteWidth / 2;
-        uly -= noteHeight / 2;
-
-        zone->SetUlx(ulx + offsetX);
-        zone->SetUly(uly + noteHeight);
-        zone->SetLrx(ulx + noteWidth + offsetX);
-        zone->SetLry(uly + noteHeight * 2);
+        zone->SetUlx(ulx - noteWidth * 0.25);
+        zone->SetUly(uly);
+        zone->SetLrx(ulx + noteWidth * 0.75);
+        zone->SetLry(uly + noteHeight);
         layer->ReorderByXPos();
         if (!AdjustPitchFromPosition(custos)) {
             LogError("Failed to set pitch.");
@@ -1649,7 +1705,7 @@ bool EditorToolkitNeume::MatchHeight(std::string elementId)
     // int itLrx;
     int offsetY;
     // int rightMost = -1;
-    double theta = staffParent->GetFacsimileInterface()->GetZone()->GetRotate();
+    double theta = (dynamic_cast<Staff *>(staffParent))->GetDrawingRotation();
 
     for (auto it = syls.begin(); it != syls.end(); ++it) {
         syl = dynamic_cast<Syl *>(*it);
@@ -1965,11 +2021,11 @@ bool EditorToolkitNeume::SetClef(std::string elementId, std::string shape)
 
     if (shape == "C") {
         clefShape = CLEFSHAPE_C;
-        shift = -3;
+        shift = 4;
     }
     else if (shape == "F") {
         clefShape = CLEFSHAPE_F;
-        shift = 3;
+        shift = -4;
     }
 
     if (clef->GetShape() != clefShape) {
@@ -2044,6 +2100,8 @@ bool EditorToolkitNeume::SetLiquescent(std::string elementId, std::string curve)
         }
     }
 
+    m_doc->GetDrawingPage()->LayOutTranscription(true);
+
     m_editInfo.import("status", "OK");
     m_editInfo.import("message", "");
     return true;
@@ -2115,10 +2173,10 @@ bool EditorToolkitNeume::Split(std::string elementId, int x)
     }
 
     // Resize current staff and insert new one filling remaining area.
+    double theta = staff->GetDrawingRotation();
     int newUlx = x;
     int newLrx = staff->GetZone()->GetLrx();
-    int newUly = staff->GetZone()->GetUly()
-        - ((x - staff->GetZone()->GetUlx()) * tan(staff->GetZone()->GetRotate() * M_PI / 180.0));
+    int newUly = staff->GetZone()->GetUly() - ((x - staff->GetZone()->GetUlx()) * tan(theta * M_PI / 180.0));
     int newLry = staff->GetZone()->GetLry(); // don't need to maintain height since we're setting rotate manually
     std::vector<std::pair<std::string, std::string>> v;
 
@@ -2140,12 +2198,11 @@ bool EditorToolkitNeume::Split(std::string elementId, int x)
         return false;
     }
 
-    splitStaff->GetZone()->SetRotate(staff->GetZone()->GetRotate());
+    splitStaff->GetZone()->SetRotate(theta);
 
     staff->GetZone()->SetLrx(x);
-    if (staff->GetZone()->GetRotate() != 0) {
-        staff->GetZone()->SetLry(
-            staff->GetZone()->GetLry() + (newLrx - x) * tan(staff->GetZone()->GetRotate() * M_PI / 180.0));
+    if (theta) {
+        staff->GetZone()->SetLry(staff->GetZone()->GetLry() + (newLrx - x) * tan(theta * M_PI / 180.0));
     }
 
     Layer *layer = vrv_cast<Layer *>(staff->GetFirst(LAYER));
@@ -2217,12 +2274,14 @@ void EditorToolkitNeume::UnlinkSyllable(Syllable *syllable)
             // Create default bounding box if facs
             if (m_doc->HasFacsimile()) {
                 Zone *zone = new Zone();
+                Staff *staff = linkedSyllable->GetAncestorStaff();
+                const int staffSize = m_doc->GetDrawingDoubleUnit(staff->m_drawingStaffSize);
 
                 zone->SetUlx(
                     linkedSyllable->GetFirst(NEUME)->GetFirst(NC)->GetFacsimileInterface()->GetZone()->GetUlx());
-                zone->SetUly(linkedSyllable->GetAncestorStaff()->GetFacsimileInterface()->GetZone()->GetLry());
+                zone->SetUly(staff->GetFacsimileInterface()->GetZone()->GetLry());
                 zone->SetLrx(linkedSyllable->GetLast(NEUME)->GetLast(NC)->GetFacsimileInterface()->GetZone()->GetLrx());
-                zone->SetLry(zone->GetUly() + 100);
+                zone->SetLry(zone->GetUly() + staffSize * 2);
 
                 // Make bbox larger if it has less than 2 ncs
                 if (linkedSyllable->GetChildCount(NC, 2) <= 2) {
@@ -3868,6 +3927,15 @@ bool EditorToolkitNeume::ChangeStaffTo(std::string elementId, std::string staffI
     return true;
 }
 
+bool EditorToolkitNeume::ParseAddSylAction(jsonxx::Object param, std::string *elementId, std::string *sylText)
+{
+    if (!param.has<jsonxx::String>("elementId")) return false;
+    (*elementId) = param.get<jsonxx::String>("elementId");
+    if (!param.has<jsonxx::String>("sylText")) return false;
+    (*sylText) = param.get<jsonxx::String>("sylText");
+    return true;
+}
+
 bool EditorToolkitNeume::ParseDragAction(jsonxx::Object param, std::string *elementId, int *x, int *y)
 {
     if (!param.has<jsonxx::String>("elementId")) return false;
@@ -4220,12 +4288,9 @@ bool EditorToolkitNeume::AdjustPitchFromPosition(Object *obj)
                 LogError("Clef is missing its parent staff.");
                 return false;
             }
-            clefOffset
-                = round((double)(clefStaff->GetDrawingY()
-                            - clefStaff->GetDrawingRotationOffsetFor(m_view->ToLogicalX(clef->GetZone()->GetUlx()))
-                            - m_view->ToLogicalY(clef->GetZone()->GetUly()))
-                    + staffSize);
         }
+        clefOffset = round((double)((staff->m_drawingLines - clef->GetLine() + 0.5) * staffSize));
+
         return true;
     };
 
@@ -4332,7 +4397,7 @@ bool EditorToolkitNeume::AdjustClefLineFromPosition(Clef *clef, Staff *staff)
 
     const double staffSize = m_doc->GetDrawingDoubleUnit(staff->m_drawingStaffSize);
     const double yDiff = clef->GetZone()->GetUly() - staff->GetZone()->GetUly()
-        + (clef->GetZone()->GetUlx() - staff->GetZone()->GetUlx()) * tan(staff->GetDrawingRotate() * M_PI / 180.0);
+        + (clef->GetZone()->GetUlx() - staff->GetZone()->GetUlx()) * tan(staff->GetDrawingRotation() * M_PI / 180.0);
     const int clefLine = staff->m_drawingLines - round(yDiff / staffSize);
     clef->SetLine(clefLine);
     return true;