+++ src/measure.cpp
@@ -391,7 +391,7 @@ void Measure::SetDrawingScoreDef(ScoreDef *drawingScoreDef)
     assert(!m_drawingScoreDef); // We should always call ResetDrawingScoreDef before
 
     m_drawingScoreDef = new ScoreDef();
-    *m_drawingScoreDef = *drawingScoreDef;
+    m_drawingScoreDef->ReplaceWithCopyOf(drawingScoreDef);
 }
 
 void Measure::ResetDrawingScoreDef()